refactor log recovery
